requirements.txt
fastapi
uvicorn
pydantic
openai
sqlmodel
asyncpg
python-dotenv